move from_repo_id to hf.rs
